src/visitors/symbolic_printer_visitor.py
# MIT License
# 
# Copyright (c) 2022 Nicola Dardanis, Lucas Weitzendorf
# 
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
# 
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
# 
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.


from src.operators.boolean_theory import BooleanConstant
from src.operators.integer_theory import IntegerConstant
from src.operators.real_theory import RealConstant
from src.operators.string_theory import StringLiteral
from src.visitors.printer_visitor import PrinterVisitor


class SymbolicPrinterVisitor(PrinterVisitor):
    def visit_boolean_constant(self, operator: BooleanConstant):
        return operator.name

    def visit_integer_constant(self, operator: IntegerConstant):
        return operator.name

    def visit_real_constant(self, operator: RealConstant):
        return operator.name

    def visit_string_literal(self, operator: StringLiteral):
        return operator.name
